html/canvas/offscreen/size.attributes.parse.empty-expected.txt

This is a testharness.js-based test.
FAIL Parsing of non-negative integers assert_equals: offscreenCanvas.width === 100 (got 0[number], expected 100[number]) expected 100 but got 0
Harness: the test ran to completion.
